cleanup + remove frame kind
