+++ crates/forge/src/cmd/test/mod.rs
@@ -347,7 +347,7 @@ impl TestArgs {
 
         // Prepare the test builder.
         let config = Arc::new(config);
-        let mut runner = MultiContractRunnerBuilder::new(config.clone())
+        let runner = MultiContractRunnerBuilder::new(config.clone())
             .set_debug(should_debug)
             .set_decode_internal(decode_internal)
             .initial_balance(evm_opts.initial_balance)
@@ -357,10 +357,8 @@ impl TestArgs {
             .enable_isolation(evm_opts.isolate)
             .fail_fast(self.fail_fast)
             .odyssey(evm_opts.odyssey)
+            .with_verbosity(verbosity)
             .build::<MultiCompiler>(project_root, &output, env, evm_opts)?;
-
-        // Set verbosity for backtrace support
-        runner.set_verbosity(verbosity);
 
         let libraries = runner.libraries.clone();
         let mut outcome = self.run_tests(runner, config, verbosity, &filter, &output).await?;
@@ -591,12 +589,12 @@ impl TestArgs {
                     sh_println!("{}", result.short_result(name))?;
 
                     // Display backtrace for failed tests when verbosity >= 3
-                    if verbosity >= 3 && result.status == TestStatus::Failure {
-                        if let Some(ref backtrace) = result.backtrace {
-                            if !backtrace.is_empty() {
-                                sh_println!("{}", backtrace)?;
-                            }
-                        }
+                    if verbosity >= 3
+                        && result.status == TestStatus::Failure
+                        && let Some(ref backtrace) = result.backtrace
+                        && !backtrace.is_empty()
+                    {
+                        sh_println!("{}", backtrace)?;
                     }
 
                     // Display invariant metrics if invariant kind.
